feat(game): let each level use the gates already built, and name levels after their gate
