+++ apps/game/src/game/__tests__/storage.test.ts
@@ -69,19 +69,19 @@ describe('drafts', () => {
   it('stores per level and leaves the others alone', () => {
     installStorage();
     writeDraft('first-wire', 'one');
-    writeDraft('not-from-nand', 'two');
-    expect(readDrafts()).toEqual({ 'first-wire': 'one', 'not-from-nand': 'two' });
+    writeDraft('not', 'two');
+    expect(readDrafts()).toEqual({ 'first-wire': 'one', 'not': 'two' });
 
     writeDraft('first-wire', 'edited');
-    expect(readDrafts()).toEqual({ 'first-wire': 'edited', 'not-from-nand': 'two' });
+    expect(readDrafts()).toEqual({ 'first-wire': 'edited', 'not': 'two' });
   });
 
   it('clears one level without disturbing the rest', () => {
     installStorage();
     writeDraft('first-wire', 'one');
-    writeDraft('not-from-nand', 'two');
+    writeDraft('not', 'two');
     clearDraft('first-wire');
-    expect(readDrafts()).toEqual({ 'not-from-nand': 'two' });
+    expect(readDrafts()).toEqual({ 'not': 'two' });
   });
 
   it('clearing a level that was never drafted is harmless', () => {
@@ -102,11 +102,11 @@ describe('progress', () => {
   it('records a gate count per level, latest pass winning', () => {
     installStorage();
     writeProgress('first-wire', { gates: 1 });
-    writeProgress('and-from-nand', { gates: 3 });
-    expect(readProgress()).toEqual({ 'first-wire': { gates: 1 }, 'and-from-nand': { gates: 3 } });
+    writeProgress('and', { gates: 3 });
+    expect(readProgress()).toEqual({ 'first-wire': { gates: 1 }, 'and': { gates: 3 } });
 
-    writeProgress('and-from-nand', { gates: 2 });
-    expect(readProgress()['and-from-nand']).toEqual({ gates: 2 });
+    writeProgress('and', { gates: 2 });
+    expect(readProgress()['and']).toEqual({ gates: 2 });
   });
 
   it('is a separate record from drafts', () => {
@@ -127,7 +127,7 @@ describe('progress', () => {
   it('yields the solved set the map reads', () => {
     installStorage();
     writeProgress('first-wire', { gates: 1 });
-    writeProgress('not-from-nand', { gates: 1 });
-    expect(new Set(Object.keys(readProgress()))).toEqual(new Set(['first-wire', 'not-from-nand']));
+    writeProgress('not', { gates: 1 });
+    expect(new Set(Object.keys(readProgress()))).toEqual(new Set(['first-wire', 'not']));
   });
 });
