+++ apps/game/src/game/solutions/not.ts
@@ -1,4 +1,4 @@
-/** Reference solution for `not-from-nand`. Proven by `__tests__/levels.test.ts`. */
+/** Reference solution for `not`. Proven by `__tests__/levels.test.ts`. */
 
 import { circuit } from '@simten/core/circuit';
 import { Led, Nand, Switch } from '@simten/core/std';
