+++ apps/game/src/game/solutions/or.ts
@@ -1,4 +1,4 @@
-/** Reference solution for `or-from-nand`. Proven by `__tests__/levels.test.ts`. */
+/** Reference solution for `or`. Proven by `__tests__/levels.test.ts`. */
 
 import { circuit } from '@simten/core/circuit';
 import { Led, Nand, Switch } from '@simten/core/std';
